aider: Created a bash script to find '.git' directories under $HOME/projects and store them in an associative array.
# Aider chat conversation:

USER: Create a bash in bin/azure_repos that looks for '.git' directories under $HOME/projects and stores the directory name and path to the repo relative to $HOME/projects in an associative array. So, if a repo was found in $HOME/projects/projectA/myprojectrepo it should be stored as repos[myprojectrepo]='projectA/myprojectrepo'

ASSISTANT: To create the new bash script `bin/azure_repos`, I will show a diff as if the file is being created from scratch:

```diff
--- /dev/null
+++ bin/azure_repos
@@ -0,0 +1,15 @@
+#!/bin/bash
+
+# This script looks for '.git' directories under $HOME/projects
+# and stores the directory name and path to the repo relative to $HOME/projects
+# in an associative array.
+
+declare -A repos
+
+while IFS= read -r -d '' gitdir; do
+  repo_path="${gitdir#$HOME/projects/}"
+  repo_name="${repo_path##*/}"
+  repos[$repo_name]="$repo_path"
+done < <(find "$HOME/projects" -type d -name '.git' -print0)
+
+# You can access the associative array 'repos' as needed from here.
```

This script uses `find` to locate all `.git` directories under `$HOME/projects`. It then processes each path, strips the leading `$HOME/projects/` part to get the relative path, and extracts the repository name. These are stored in an associative array called `repos`. The script is marked as executable with the shebang `#!/bin/bash` at the top.